# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 'StockBeat API Documentation'
  description: 'Token-authenticated REST API consumed by the StockBeat mobile app (React Native).'
  version: 1.0.0
servers:
  - url: 'https://stockbeat.ddev.site'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain a token via `POST /auth/otp/request` then `POST /auth/otp/verify` (Laravel Sanctum bearer token).'
security:
  -
    default: []
paths:
...
  /api/v1/settings/notifications:
    get:
      summary: 'Get notification preferences.'
      operationId: getNotificationPreferences
      description: 'Returns defaults (all channels enabled, no quiet hours) if the user has never saved any.'
      parameters: []
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: null
                  data:
                    preferences:
                      push_enabled: true
                      email_enabled: true
                      sms_enabled: false
                      quiet_hours_start: '21:00'
                      quiet_hours_end: '07:00'
                      quiet_hours_timezone: Australia/Sydney
                      sound: default
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    type: object
                    properties:
                      preferences:
                        type: object
                        properties:
                          push_enabled:
                            type: boolean
                            example: true
                          email_enabled:
                            type: boolean
                            example: true
                          sms_enabled:
                            type: boolean
                            example: false
                          quiet_hours_start:
                            type: string
                            example: '21:00'
                          quiet_hours_end:
                            type: string
                            example: '07:00'
                          quiet_hours_timezone:
                            type: string
                            example: Australia/Sydney
                          sound:
                            type: string
                            example: default
      tags:
        - Settings
    put:
      summary: 'Update notification preferences.'
      operationId: updateNotificationPreferences
      description: ''
      parameters: []
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: null
                  data:
                    preferences:
                      push_enabled: true
                      email_enabled: false
                      sms_enabled: false
                      quiet_hours_start: '21:00'
                      quiet_hours_end: '07:00'
                      quiet_hours_timezone: Australia/Sydney
                      sound: chime
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: null
                    nullable: true
                  data:
                    type: object
                    properties:
                      preferences:
                        type: object
                        properties:
                          push_enabled:
                            type: boolean
                            example: true
                          email_enabled:
                            type: boolean
                            example: false
                          sms_enabled:
                            type: boolean
                            example: false
                          quiet_hours_start:
                            type: string
                            example: '21:00'
                          quiet_hours_end:
                            type: string
                            example: '07:00'
                          quiet_hours_timezone:
                            type: string
                            example: Australia/Sydney
                          sound:
                            type: string
                            example: chime
      tags:
        - Settings
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                push_enabled:
                  type: boolean
                  description: ''
                  example: false
                email_enabled:
                  type: boolean
                  description: ''
                  example: false
                sms_enabled:
                  type: boolean
                  description: ''
                  example: true
                quiet_hours_start:
                  type: string
                  description: 'Must be a valid date in the format <code>H:i</code>.'
                  example: '18:25'
                  nullable: true
                quiet_hours_end:
                  type: string
                  description: 'Must be a valid date in the format <code>H:i</code>.'
                  example: '18:25'
                  nullable: true
                quiet_hours_timezone:
                  type: string
                  description: 'Must be a valid time zone, such as <code>Africa/Accra</code>.'
                  example: Asia/Yekaterinburg
                  nullable: true
                sound:
                  type: string
                  description: 'Must not be greater than 50 characters.'
                  example: 'n'
...